+++ index.js
@@ -13,6 +13,14 @@
 var accepts = require('accepts')
 var escapeHtml = require('escape-html');
 var fs = require('fs');
+var util = require('util')
+
+/**
+ * Module variables.
+ */
+
+var inspect = util.inspect
+var toString = Object.prototype.toString
 
 /**
  * Error handler:
@@ -56,7 +64,7 @@ exports = module.exports = function errorHandler(){
 
     // write error to console
     if (env !== 'test') {
-      console.error(err.stack || String(err))
+      console.error(stringify(err))
     }
 
     // cannot actually respond
@@ -77,7 +85,7 @@ exports = module.exports = function errorHandler(){
         if (e) return next(e);
         fs.readFile(__dirname + '/public/error.html', 'utf8', function(e, html){
           if (e) return next(e);
-          var stack = (err.stack || '')
+          var stack = String(err.stack || '')
             .split('\n').slice(1)
             .map(function(v){ return '<li>' + escapeHtml(v).replace(/  /g, ' &nbsp;') + '</li>'; }).join('');
             html = html
@@ -85,7 +93,7 @@ exports = module.exports = function errorHandler(){
               .replace('{stack}', stack)
               .replace('{title}', escapeHtml(exports.title))
               .replace('{statusCode}', res.statusCode)
-              .replace(/\{error\}/g, escapeHtml(String(err)).replace(/  /g, ' &nbsp;').replace(/\n/g, '<br>'));
+              .replace(/\{error\}/g, escapeHtml(stringify(err)).replace(/  /g, ' &nbsp;').replace(/\n/g, '<br>'));
             res.setHeader('Content-Type', 'text/html; charset=utf-8');
             res.end(html);
         });
@@ -100,7 +108,7 @@ exports = module.exports = function errorHandler(){
     // plain text
     } else {
       res.setHeader('Content-Type', 'text/plain');
-      res.end(err.stack || String(err));
+      res.end(stringify(err));
     }
   };
 };
@@ -110,3 +118,22 @@ exports = module.exports = function errorHandler(){
  */
 
 exports.title = 'Connect';
+
+/**
+ * Stringify a value.
+ * @api private
+ */
+
+function stringify(val) {
+  var stack = val.stack
+
+  if (stack) {
+    return String(stack)
+  }
+
+  var str = String(val)
+
+  return str === toString.call(val)
+    ? inspect(val)
+    : str
+}